examples/requirements.txt
cryptography
pyjwt
requests
flask
